Create Login view and isolate WebSocket functionality
